+++ src/clj_watson/cli.clj
@@ -1,10 +1,15 @@
 (ns clj-watson.cli
   (:gen-class)
   (:require
-   [clj-watson.cli-spec :as cli-spec]
    [clj-watson.entrypoint :as entrypoint]))
+
+;; mock/capture target
+(defn system-exit [code]
+  (System/exit code))
 
 (defn -main
   "Entrypoint for -M cli usage"
   [& args]
-  (entrypoint/do-scan (cli-spec/parse-args args)))
+  (let [{:keys [exit]} (entrypoint/scan-main args)]
+    (when exit
+      (system-exit exit))))